moved payload checing into a utilties module
